requirements/base.txt
Django
django-environ
#psycopg2==2.9.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

celery==5.4.0
django-celery-results
django-celery-beat

whitenoise==6.9.0

django-filter==25.1
#django-extensions==3.2.3
django-cors-headers==4.7.0
django-storages==1.14.5

python-decouple
redis

#scikit-surprise
jupyter
numpy
pandas


amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0

click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6

kombu==5.3.7
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13

boto3==1.37.13
attrs==25.3.0

gunicorn==23.0.0
# sentry-sdk==2.22.0

# requests==2.32.3

# google-api-python-client==2.164.0
# google-auth==2.38.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.1

# pyotp==2.9.0
# qrcode==8.0
# tabulate==0.9.0